requirements.txt
# Google Agent Development Kit
google-adk

# Additional dependencies that might be needed
# Uncomment if you encounter import errors
fastapi
uvicorn
# python-multipart